requirements.txt
numpy~=1.26
pandas~=2.2
scikit-learn~=1.5
networkx~=3.3
langchain>=0.3
langchain-openai>=0.2
langchain-anthropic>=0.3
langchain-community>=0.3
langchain-core>=0.3
tenacity>=8.2
langgraph>=1.0
langgraph-checkpoint-sqlite>=1.0
langsmith>=0.1
openai~=1.40
scipy~=1.13
python-dotenv~=1.0
faiss-cpu~=1.8
pyyaml~=6.0
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29
psycopg[binary]>=3.1
alembic>=1.13
langgraph-checkpoint-postgres>=1.0
pgvector>=0.3
typing-extensions~=4.12
streamlit>=1.35
plotly>=5.20
fastapi>=0.111
uvicorn[standard]>=0.29
black~=24.0
ruff~=0.5
pre-commit~=3.7